requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Flask==3.0.3
google-ai-generativelanguage==0.6.5
google-api-core==1.34.1
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.48.2
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
Hypercorn==0.17.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
opencv-python-headless==4.10.0.84
packaging==24.1
# pdf2image==1.17.0
# poppler-utils==0.1.0
priority==2.0.0
proto-plus==1.23.0
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pyOpenSSL==24.1.0
pyparsing==3.1.2
requests==2.32.3
rsa==4.9
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
tqdm==4.66.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
wsproto==1.2.0
zope.interface==6.4.post2
# python-dateutil
PyMuPDF==1.24.5
# fitz==0.0.1.dev2